+++ molo/core/views.py
@@ -1,3 +1,6 @@
+import pkg_resources
+import requests
+
 from django.contrib import messages
 from django.core.urlresolvers import reverse
 from django.http import HttpResponse
@@ -9,6 +12,7 @@ from molo.core.utils import generate_slug
 from molo.core.models import PageTranslation, SiteLanguage
 
 from wagtail.wagtailcore.models import Page
+from molo.core.known_plugins import known_plugins
 
 
 def locale_set(request, locale):
@@ -57,3 +61,39 @@ def add_translation(request, page_id, locale):
 
 def import_from_git(request):
     return render(request, 'admin/import_from_git.html')
+
+
+def versions(request):
+    comparison_url = "https://github.com/praekelt/%s/compare/%s...%s"
+    plugins_info = []
+    for plugin in known_plugins():
+        try:
+            plugin_version = (
+                pkg_resources.get_distribution(plugin[0])).version
+
+            pypi_version = get_pypi_version(plugin[0])
+
+            if plugin[0] == 'molo.core':
+                compare_versions_link = comparison_url % (
+                    'molo', plugin_version, pypi_version)
+            else:
+                compare_versions_link = comparison_url % (
+                    plugin[0], plugin_version, pypi_version)
+
+            plugins_info.append((plugin[1], pypi_version,
+                                 plugin_version, compare_versions_link))
+        except:
+            pypi_version = get_pypi_version(plugin[0])
+            plugins_info.append((plugin[1], pypi_version, "-",
+                                 ""))
+
+    return render(request, 'admin/versions.html', {
+        'plugins_info': plugins_info,
+    })
+
+
+
+def get_pypi_version(plugin_name):
+    url = "https://pypi.python.org/pypi/%s/json"
+    content = requests.get(url % plugin_name).json()
+    return content.get('info').get('version')